+++ client/src/App.js
@@ -5,6 +5,9 @@ import ApolloClient from "apollo-boost";
 import SearchGames from "./pages/SearchGames";
 import SavedGames from "./pages/SavedGames";
 import Navbar from "./components/Navbar";
+//import SignupForm from "./components/SignupForm";
+import Signup from "./pages/Signup";
+import Login from "./pages/Login";
 
 const client = new ApolloClient({
   request: (operation) => {
@@ -28,6 +31,8 @@ function App() {
           <Switch>
             <Route exact path="/" component={SearchGames} />
             <Route exact path="/saved" component={SavedGames} />
+            <Route exact path="/signup" component={Signup} />
+            <Route exact path="/login" component={Login} />
             <Route render={() => <h1 className="display-2">Wrong page!</h1>} />
           </Switch>
         </>
