--- client/src/components/LoginForm.js
@@ -1,95 +1,0 @@
-// see SignupForm.js for comments
-import React, { useState } from "react";
-import Auth from "../utils/auth";
-import { Form, Button, Alert } from "react-bootstrap";
-import { LOGIN_USER } from "../utils/mutations";
-import { useMutation } from "@apollo/react-hooks";
-
-const LoginForm = () => {
-  const [login, { error }] = useMutation(LOGIN_USER);
-  const [userFormData, setUserFormData] = useState({ email: "", password: "" });
-  const [validated] = useState(false);
-  const [showAlert, setShowAlert] = useState(false);
-
-  const handleInputChange = (event) => {
-    const { name, value } = event.target;
-    setUserFormData({
-      ...userFormData,
-      [name]: value,
-    });
-  };
-
-  const handleFormSubmit = async (event) => {
-    event.preventDefault();
-
-    try {
-      const { data } = await login({
-        variables: { ...userFormData },
-      });
-      Auth.login(data.login.token);
-    } catch (e) {
-      console.error(e);
-      setShowAlert(true);
-    }
-
-    setUserFormData({
-      username: "",
-      email: "",
-      password: "",
-    });
-  };
-
-  return (
-    <>
-      <Form noValidate validated={validated} onSubmit={handleFormSubmit}>
-        <Alert
-          dismissible
-          onClose={() => setShowAlert(false)}
-          show={showAlert}
-          variant="danger"
-        >
-          Something went wrong with your login credentials!
-        </Alert>
-        <Form.Group>
-          <Form.Label htmlFor="email">Email</Form.Label>
-          <Form.Control
-            type="text"
-            placeholder="Your email"
-            name="email"
-            onChange={handleInputChange}
-            value={userFormData.email}
-            required
-          />
-          <Form.Control.Feedback type="invalid">
-            Email is required!
-          </Form.Control.Feedback>
-        </Form.Group>
-
-        <Form.Group>
-          <Form.Label htmlFor="password">Password</Form.Label>
-          <Form.Control
-            type="password"
-            placeholder="Your password"
-            name="password"
-            onChange={handleInputChange}
-            value={userFormData.password}
-            required
-          />
-          <Form.Control.Feedback type="invalid">
-            Password is required!
-          </Form.Control.Feedback>
-        </Form.Group>
-        <Button
-          disabled={!(userFormData.email && userFormData.password)}
-          type="submit"
-          variant="success"
-        >
-          Submit
-        </Button>
-        {error && <div>Login failed</div>}
-      </Form>
-    </>
-  );
-};
-
-export default LoginForm;
