+++ client/src/pages/SearchGames.js
@@ -10,7 +10,7 @@ import {
 } from "react-bootstrap";
 
 import Auth from "../utils/auth";
-import { searchIGDB } from "../utils/API";
+// import { searchIGDB } from "../utils/API";
 import { saveGameIds, getSavedGameIds } from "../utils/localStorage";
 import { useMutation } from "@apollo/react-hooks";
 import { SAVE_GAME } from "../utils/mutations";
@@ -35,34 +35,34 @@ const SearchGames = () => {
   });
 
   // create method to search for games and set state on form submit
-  const handleFormSubmit = async (event) => {
-    event.preventDefault();
+  // const handleFormSubmit = async (event) => {
+  //   event.preventDefault();
 
-    if (!searchInput) {
-      return false;
-    }
+  //   if (!searchInput) {
+  //     return false;
+  //   }
 
-    try {
-      const response = await searchIGDB(searchInput);
+  //   try {
+  //     const response = await searchIGDB(searchInput);
 
-      if (!response.ok) {
-        throw new Error("something went wrong!");
-      }
+  //     if (!response.ok) {
+  //       throw new Error("something went wrong!");
+  //     }
 
-      const { items } = await response.json();
+  //     const { items } = await response.json();
 
-      const gameData = items.map((game) => ({
-        gameId: game.id,
-        gameName: game.name,
-        cover: game.cover.url,
-      }));
+  //     const gameData = items.map((game) => ({
+  //       gameId: game.id,
+  //       gameName: game.name,
+  //       cover: game.cover.url,
+  //     }));
 
-      setSearchedGames(gameData);
-      setSearchInput("");
-    } catch (err) {
-      console.error(err);
-    }
-  };
+  //     setSearchedGames(gameData);
+  //     setSearchInput("");
+  //   } catch (err) {
+  //     console.error(err);
+  //   }
+  // };
 
   // create function to handle saving a game to our database
   const handleSaveGame = async (gameId) => {
@@ -102,7 +102,7 @@ const SearchGames = () => {
       <Jumbotron fluid className="text-light bg-dark">
         <Container>
           <h1>Search for Games!</h1>
-          <Form onSubmit={handleFormSubmit}>
+          {/* <Form onSubmit={handleFormSubmit}>
             <Form.Row>
               <Col xs={12} md={8}>
                 <Form.Control
@@ -120,7 +120,7 @@ const SearchGames = () => {
                 </Button>
               </Col>
             </Form.Row>
-          </Form>
+          </Form> */}
         </Container>
       </Jumbotron>
