+++ client/src/pages/Signup.js
@@ -1,0 +1,121 @@
+import React, { useState } from "react";
+import { Form, Button, Alert } from "react-bootstrap";
+import { ADD_USER } from "../utils/mutations";
+import { useMutation } from "@apollo/react-hooks";
+import Auth from "../utils/auth";
+
+const Signup = () => {
+  const [addUser, { error }] = useMutation(ADD_USER);
+  // set initial form state
+  const [userFormData, setUserFormData] = useState({
+    username: "",
+    email: "",
+    password: "",
+  });
+  // set state for form validation
+  const [validated] = useState(false);
+  // set state for alert
+  const [showAlert, setShowAlert] = useState(false);
+
+  const handleInputChange = (event) => {
+    const { name, value } = event.target;
+    setUserFormData({ ...userFormData, [name]: value });
+  };
+
+  const handleFormSubmit = async (event) => {
+    event.preventDefault();
+
+    // check if form has everything (as per react-bootstrap docs)
+    // const form = event.currentTarget;
+    // if (form.checkValidity() === false) {
+    //   event.preventDefault();
+    //   event.stopPropagation();
+    // }
+
+    try {
+      const { data } = await addUser({
+        variables: { ...userFormData },
+      });
+
+      Auth.login(data.addUser.token);
+    } catch (e) {
+      console.error(e);
+      setShowAlert(true);
+    }
+
+    setUserFormData({
+      username: "",
+      email: "",
+      password: "",
+    });
+  };
+
+  return (
+    <Form noValidate validated={validated} onSubmit={handleFormSubmit}>
+      {/* show alert if server response is bad */}
+      <Alert
+        dismissible
+        onClose={() => setShowAlert(false)}
+        show={showAlert}
+        variant="danger"
+      >
+        Something went wrong with your signup!
+      </Alert>{" "}
+      <Form.Group>
+        <Form.Label htmlFor="username">Username</Form.Label>
+        <Form.Control
+          type="text"
+          placeholder="Your username"
+          name="username"
+          onChange={handleInputChange}
+          value={userFormData.username}
+          required
+        />
+      </Form.Group>
+      <Form.Group>
+        <Form.Label htmlFor="email">Email</Form.Label>
+        <Form.Control
+          type="email"
+          placeholder="Your email address"
+          name="email"
+          onChange={handleInputChange}
+          value={userFormData.email}
+          required
+        />
+        <Form.Control.Feedback type="invalid">
+          Email is required!
+        </Form.Control.Feedback>
+      </Form.Group>
+      <Form.Group>
+        <Form.Label htmlFor="password">Password</Form.Label>
+        <Form.Control
+          type="password"
+          placeholder="Your password"
+          name="password"
+          onChange={handleInputChange}
+          value={userFormData.password}
+          required
+        />
+        <Form.Control.Feedback type="invalid">
+          Password is required!
+        </Form.Control.Feedback>
+      </Form.Group>
+      <Button
+        disabled={
+          !(
+            userFormData.username &&
+            userFormData.email &&
+            userFormData.password
+          )
+        }
+        type="submit"
+        variant="success"
+      >
+        Submit
+      </Button>
+      {error && <div>Sign up failed</div>}
+    </Form>
+  );
+};
+
+export default Signup;
